+++ app/api/ada/lib/storage/bridge/tests/multiwallet.test.js
@@ -24,10 +24,8 @@ import {
 import {
   asGetAllUtxos,
   asDisplayCutoff,
-} from '../../models/Bip44Wallet/traits';
-import {
   asGetUtxoBalance,
-} from '../../models/common/traits';
+} from '../../models/PublicDeriver/traits';
 
 import {
   updateTransactions
