+++ app/api/ada/lib/storage/models/common/interfaces.js
@@ -1,9 +1,6 @@
 // @flow
 
 import type { lf$Transaction } from 'lovefield';
-import {
-  BigNumber
-} from 'bignumber.js';
 
 import type {
   KeyRow,
@@ -26,28 +23,6 @@ export type RawTableVariation<Func, Deps, Arg> = (
   tableMap: Map<number, string>,
 ) => ReturnType<Func>;
 
-export type Address = {|
-  +address: string,
-|};
-export type Value = {|
-  /**
-   * note: an undefined value is different than a value of 0
-   * since you can have a UTXO with a value of 0
-   * which is different from having no UTXO at all
-   */
-  +value: void | BigNumber,
-|};
-export type Addressing = {|
-  +addressing: {|
-    +path: Array<number>,
-    +startLevel: number,
-  |}
-|};
-
-export type UsedStatus = {|
-  isUsed: boolean,
-|};
-
 export type IChangePasswordRequest = {
   oldPassword: null | string,
   newPassword: null | string,
@@ -67,13 +42,4 @@ export type IRenameFunc = (
 ) => Promise<IRenameResponse>;
 export interface IRename {
   +rename: IRenameFunc;
-}
-
-export type IGetBalanceRequest = void;
-export type IGetBalanceResponse = BigNumber;
-export type IGetBalanceFunc = (
-  body: IGetBalanceRequest
-) => Promise<IGetBalanceResponse>;
-export interface IGetBalance {
-  +getBalance: IGetBalanceFunc;
 }
