+++ app/api/ada/transactions/shelley/utxoTransactions.js
@@ -21,8 +21,8 @@ import {
 } from '../../lib/storage/database/walletTypes/bip44/api/utils';
 import type {
   Address, Value, Addressing,
-} from '../../lib/storage/models/common/interfaces';
-import type { IGetAllUtxosResponse } from '../../lib/storage/models/PublicDeriver/interfaces';
+  IGetAllUtxosResponse
+} from '../../lib/storage/models/PublicDeriver/interfaces';
 import { v2SkKeyToV3Key } from '../utils';
 
 declare var CONFIG: ConfigType;
